+++ app/components/Views/confirmations/Confirm/Confirm.test.tsx
@@ -115,7 +115,7 @@ describe('Confirm', () => {
     expect(getByTestId('flat-confirmation-container')).toBeDefined();
   });
 
-  it('renders correct information for personal sign', () => {
+  it('renders information for personal sign', () => {
     const { getAllByRole, getByText } = renderWithProvider(
       <SafeAreaProvider>
         <Confirm />
@@ -135,7 +135,7 @@ describe('Confirm', () => {
     expect(getAllByRole('button')).toHaveLength(2);
   });
 
-  it('renders correct information for typed sign v1', () => {
+  it('renders information for typed sign v1', () => {
     const { getAllByRole, getAllByText, getByText, queryByText } =
       renderWithProvider(
         <SafeAreaProvider>
@@ -154,7 +154,7 @@ describe('Confirm', () => {
     expect(queryByText('This is a deceptive request')).toBeNull();
   });
 
-  it('renders correct information for staking deposit', async () => {
+  it('renders information for staking deposit', async () => {
     const { getByText } = renderWithProvider(<Confirm />, {
       state: stakingDepositConfirmationState,
     });
@@ -166,7 +166,7 @@ describe('Confirm', () => {
     expect(getByText('Advanced details')).toBeDefined();
   });
 
-  it('renders correct information for staking withdrawal', async () => {
+  it('renders information for staking withdrawal', async () => {
     const { getByText } = renderWithProvider(<Confirm />, {
       state: stakingWithdrawalConfirmationState,
     });
@@ -177,7 +177,7 @@ describe('Confirm', () => {
     expect(getByText('Network Fee')).toBeDefined();
   });
 
-  it('renders correct information for staking claim', async () => {
+  it('renders information for staking claim', async () => {
     const { getByText } = renderWithProvider(<Confirm />, {
       state: stakingClaimConfirmationState,
     });
@@ -190,7 +190,7 @@ describe('Confirm', () => {
     expect(getByText('Network Fee')).toBeDefined();
   });
 
-  it('renders correct information for contract interaction', async () => {
+  it('renders information for contract interaction', async () => {
     jest.spyOn(ConfirmationRedesignEnabled, 'useConfirmationRedesignEnabled')
       .mockReturnValue({ isRedesignedEnabled: true });
