+++ app/components/Views/confirmations/components/Confirm/Title/Title.test.tsx
@@ -46,7 +46,7 @@ describe('Confirm Title', () => {
     ).toBeTruthy();
   });
 
-  it('should render correct title and subtitle for personal siwe request', () => {
+  it('renders correct title and subtitle for personal siwe request', () => {
     const { getByText } = renderWithProvider(<Title />, {
       state: siweSignatureConfirmationState,
     });
@@ -56,23 +56,7 @@ describe('Confirm Title', () => {
     ).toBeTruthy();
   });
 
-  it('should render correct title and subtitle for contract interaction', () => {
-    jest.spyOn(TransactionMetadataRequestHook, 'useTransactionMetadataRequest').mockReturnValue({
-      id: '7e62bcb1-a4e9-11ef-9b51-ddf21c91a998',
-      type: TransactionType.contractInteraction,
-      txParams: {
-        data: '0x123456',
-        from: '0x935e73edb9ff52e23bac7f7e043a1ecd06d05477',
-        to: '0x1234567890123456789012345678901234567890',
-        value: '0x0',
-      },
-      chainId: '0x1' as `0x${string}`,
-      networkClientId: 'mainnet',
-      status: TransactionStatus.unapproved,
-      time: Date.now(),
-      origin: 'https://metamask.github.io',
-    });
-
+  it('renders correct title and subtitle for contract interaction', () => {
     const { getByText } = renderWithProvider(<Title />, {
       state: generateContractInteractionState,
     });
