+++ app/selectors/confirmTransaction.ts
@@ -1,8 +1,9 @@
-import { selectTransactions } from './transactionController';
-import { RootState } from '../reducers';
-import { selectGasFeeControllerEstimates } from './gasFeeController';
 import { mergeGasFeeEstimates } from '@metamask/transaction-controller';
 import { createSelector } from 'reselect';
+import { RootState } from '../reducers';
+import { getFeatureFlagValue, selectRemoteFeatureFlags } from './featureFlagController';
+import { selectGasFeeControllerEstimates } from './gasFeeController';
+import { selectTransactions } from './transactionController';
 import { createDeepEqualSelector } from './util';
 
 const selectCurrentTransactionId = (state: RootState) => state.transaction?.id;
@@ -45,5 +46,42 @@ export const selectGasFeeEstimates = createSelector(
     }
 
     return gasFeeControllerEstimates;
+  },
+);
+
+export interface ConfirmationRedesignRemoteFlags {
+  signatures: boolean;
+  staking_confirmations: boolean;
+  contract_interaction: boolean;
+}
+
+export const selectConfirmationRedesignFlags = createSelector(
+  selectRemoteFeatureFlags,
+  (remoteFeatureFlags) => {
+    const confirmationRedesignFlags =
+      (remoteFeatureFlags?.confirmation_redesign as unknown as ConfirmationRedesignRemoteFlags) ??
+      {};
+
+    const isStakingConfirmationsEnabled = getFeatureFlagValue(
+      process.env.FEATURE_FLAG_REDESIGNED_STAKING_TRANSACTIONS,
+      confirmationRedesignFlags.staking_confirmations,
+    );
+
+    const isContractInteractionEnabled = getFeatureFlagValue(
+      process.env.FEATURE_FLAG_REDESIGNED_CONTRACT_INTERACTION,
+      confirmationRedesignFlags.contract_interaction,
+    )
+
+    const isSignaturesEnabled = getFeatureFlagValue(
+      process.env.FEATURE_FLAG_REDESIGNED_SIGNATURES,
+      confirmationRedesignFlags.signatures,
+    );
+
+    return {
+      ...confirmationRedesignFlags,
+      staking_confirmations: isStakingConfirmationsEnabled,
+      signatures: isSignaturesEnabled,
+      contract_interaction: isContractInteractionEnabled,
+    };
   },
 );
