+++ app/selectors/featureFlagController/index.ts
@@ -1,12 +1,18 @@
 import { createSelector } from 'reselect';
-import { StateWithPartialEngine } from './types';
 import { isRemoteFeatureFlagOverrideActivated } from '../../core/Engine/controllers/remote-feature-flag-controller';
-import { getFeatureFlagValue } from './env';
+import { StateWithPartialEngine } from './types';
 
-export interface ConfirmationRedesignRemoteFlags {
-  signatures: boolean;
-  staking_confirmations: boolean;
-  contract_interaction: boolean;
+export function getFeatureFlagValue(
+  envValue: string | undefined,
+  remoteValue: boolean,
+): boolean {
+  if (envValue === 'true') {
+    return true;
+  }
+  if (envValue === 'false') {
+    return false;
+  }
+  return remoteValue;
 }
 
 export const selectRemoteFeatureFlagControllerState = (
@@ -22,35 +28,3 @@ export const selectRemoteFeatureFlags = createSelector(
     return remoteFeatureFlagControllerState?.remoteFeatureFlags ?? {};
   },
 );
-
-export const selectConfirmationRedesignFlags = createSelector(
-  selectRemoteFeatureFlags,
-  (remoteFeatureFlags) => {
-    const confirmationRedesignFlags =
-      (remoteFeatureFlags?.confirmation_redesign as unknown as ConfirmationRedesignRemoteFlags) ??
-      {};
-
-    const isStakingConfirmationsEnabled = getFeatureFlagValue(
-      process.env.FEATURE_FLAG_REDESIGNED_STAKING_TRANSACTIONS,
-      confirmationRedesignFlags.staking_confirmations,
-    );
-
-    const isContractInteractionEnabled = getFeatureFlagValue(
-      process.env.FEATURE_FLAG_REDESIGNED_CONTRACT_INTERACTION,
-      confirmationRedesignFlags.contract_interaction,
-    )
-
-    const isSignaturesEnabled = getFeatureFlagValue(
-      process.env.FEATURE_FLAG_REDESIGNED_SIGNATURES,
-      confirmationRedesignFlags.signatures,
-    );
-
-    return {
-      ...confirmationRedesignFlags,
-      staking_confirmations: isStakingConfirmationsEnabled,
-      signatures: isSignaturesEnabled,
-      contract_interaction: isContractInteractionEnabled,
-    };
-  },
-);
-
